Create HTMLPage, i18n files ru, en. Move select checkbox step into Action in HTMLPage.
