+++ java/page-factory-2/html-elements-plugin/src/main/resources/i18n/ru/sbtqa/tag/pagefactory/HTMLPage/en.properties
@@ -1,0 +1,1 @@
+ru.sbtqa.tag.pagefactory.select.checkBox=select checkbox
