+++ java/page-factory-2/html-elements-plugin/src/main/resources/i18n/ru/sbtqa/tag/pagefactory/htmlstepdefs/HTMLStepDefs/en.properties
@@ -4,4 +4,3 @@ ru.sbtqa.tag.pagefactory.userActionInBlockOneParam=^user in block \"([^\"]*)\" \
 ru.sbtqa.tag.pagefactory.userActionInBlockTwoParams=^user in block \"([^\"]*)\" \\((.*?)\\) with the parameters \"([^\"]*)\"  \"([^\"]*)\"$
 ru.sbtqa.tag.pagefactory.findElementInBlock=^user in block \"([^\"]*)\" finds (element|textinput|checkbox|radiobutton|table|header|button|link|image) \"([^\"]*)\"$
 ru.sbtqa.tag.pagefactory.findElementInList=^user in list \"([^\"]*)\" finds the value element \"([^\"]*)\"$
-ru.sbtqa.tag.pagefactory.select.checkBox=^user select checkbox \"([^\"]*)\"$
