+++ java/page-factory-2/html-elements-plugin/src/main/resources/i18n/ru/sbtqa/tag/pagefactory/htmlstepdefs/HTMLStepDefs/ru.properties
@@ -4,4 +4,3 @@ ru.sbtqa.tag.pagefactory.userActionInBlockOneParam=^(?:пользователь 
 ru.sbtqa.tag.pagefactory.userActionInBlockTwoParams=^(?:пользователь |он |)в блоке \"([^\"]*)\" \\((.*?)\\) с параметрами \"([^\"]*)\" \"([^\"]*)\"$
 ru.sbtqa.tag.pagefactory.findElementInBlock=^(?:пользователь |он |)в блоке \"([^\"]*)\" находит (элемент|текстовое поле|чекбокс|радиобатон|таблицу|заголовок|кнопку|ссылку|изображение) \"([^\"]*)\"$
 ru.sbtqa.tag.pagefactory.findElementInList=^(?:пользователь |он |)в списке \"([^\"]*)\" находит элемент со значением \"([^\"]*)\"$
-ru.sbtqa.tag.pagefactory.select.checkBox= ^(?:пользователь |он |)отмечает признак в чекбоксе \"([^\"]*)\"$
